More fmal() fixes

+++ include/lmms_math.h
@@ -131,7 +131,7 @@ static inline int fast_rand()
 }
 
 //! @brief Takes advantage of fmal() function if present in hardware
-static inline long fastFmal( long a, long b, long c ) {
+static inline long double fastFmal( long double a, long double b, long double c ) {
 #ifdef FP_FAST_FMAF
 	#ifdef __clang__
 		return fma( a, b, c );
@@ -157,7 +157,7 @@ static inline float fastFmaf( float a, float b, float c ) {
 }
 
 //! @brief Takes advantage of fma() function if present in hardware
-static inline int fastFma( int a, int b, int c ) {
+static inline double fastFma( double a, double b, double c ) {
 #ifdef FP_FAST_FMAF
 	return fma( a, b, c );
 #else
